backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Machine Learning and AI
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
transformers==4.40.0
sentencepiece==0.2.1
protobuf==4.21.12
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Data processing
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# HTTP client
httpx==0.25.2
requests==2.31.0

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# PDF generation
reportlab==4.0.7

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring and logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
